feat(daemon): support running a daemon that executes scripts on email received (#23)
A new command `daemon` has been added that will allow the cli to run as a daemon process. It connects to the Mailscript backend via the `daemon-bridge` over websockets. The connection is to a names daemon accessory which can be included in workflows.

* feat(accessories): allow adding 'daemon' type
* feat(daemon): update the wait for emails command

The structure has been changed to be a json object with {subject,text}.

* test(daemon): add tests to cover "parts" of daemon
* feat(daemon): setup reconnect and ping infrastructure
* feat(daemon): rework logging for reconnect